Create README.md
I recently became a Substitute Teacher for the Windward School District here in Oahu, Hawaii. I am so looking forward to meeting other teachers and learning and inspiring each other.  I have a real heart for children, especially elementary school age children.   This is why I choose  to be an Elementary School, Substitute. I love to play and have fun and enjoy teaching in an interactive, multi media manner. I look forward to meeting you and the privilege  to make a difference in our children, as their Substitute Teacher!